png2jpg

+++ png2jpg.py
@@ -11,9 +11,11 @@ for i in arr:
         lista = os.listdir(path+i+"/"+x)
         for idx,y in enumerate(lista):
             name = y.split(".")[0]
-            # Load .png image
-            image = cv2.imread(path+i+"/"+x+"/"+y)
-            # Save .jpg image
-            cv2.imwrite(path+i+"/"+x+"/"+name+'.jpg', image, [int(cv2.IMWRITE_JPEG_QUALITY), 100])
-            os.remove(path+i+"/"+x+"/"+y)
-            print(y)
+            if "png" in y: 
+                # Load .png image
+                image = cv2.imread(path+i+"/"+x+"/"+y)
+                # Save .jpg image
+                cv2.imwrite(path+i+"/"+x+"/"+name+'.jpg', image, [int(cv2.IMWRITE_JPEG_QUALITY), 100])
+                os.remove(path+i+"/"+x+"/"+y)
+                print(y)
+
